Kernel: Add unicode support

+++ kernel/include/kernel/VESA.h
@@ -26,7 +26,7 @@ namespace VESA
 	};
 
 	bool Initialize();
-	void PutEntryAt(uint8_t, uint32_t, uint32_t, Color, Color);
+	void PutEntryAt(uint16_t, uint32_t, uint32_t, Color, Color);
 	void Clear(Color);
 	void ScrollLine(uint32_t line);
